Extracted Renderable interface from SGML
